requirements.txt
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.0
scipy==1.12.0
schedule==1.2.1
APScheduler==3.10.4
